Setup an environment for the digit recognition competition
To start, run the following command in your repo:
    sudo python setup.py develop

You can then use the digit package functionnalities calling
    import digit
in your programs.

You have to download the training data (train.csv) in the digit/
directory.

To develop, place the basic functions in the digit package
and wrap it up in scripts written in the scripts/ folder.

Don't forget to comment your code with docstring for a better
reuse.